src/modules/users/services/SendForgotPasswordEmailService.ts
import { inject, injectable } from 'tsyringe';

import IUsersRepository from '../repositories/IUsersRepository';
import IUserTokensRepository from '../repositories/IUserTokensRepository';
import IEmailProvider from '../../../app/container/providers/EmailProvider/models/IEmailProvider';

import AppError from '../../../app/errors/AppError';

interface Request {
  email: string;
}

@injectable()
class SendForgotPasswordEmailService {
  constructor(
    @inject('UsersRepository')
    private usersRepository: IUsersRepository,

    @inject('EmailProvider')
    private emailProvider: IEmailProvider,

    @inject('UserTokensRepository')
    private userTokensRepository: IUserTokensRepository
  ) {}

  public async execute({ email }: Request): Promise<void> {
    const user = await this.usersRepository.findByEmail(email);

    if (!user) {
      throw new AppError('User does not exists.');
    }

    const { token } = await this.userTokensRepository.generate(user.id);

    await this.emailProvider.sendEmail(
      email,
      `Seu pedido de recuperação de senha foi recebido. \n Token de Recuperação: ${token}`
    );
  }
}

export default SendForgotPasswordEmailService;
